+++ src/components/donation/Steps/DonateMethods/Crypto/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Crypto";
+export { default } from "./Form";
